+++ README.md
@@ -39,8 +39,7 @@ look right, but you _will_ still be able to build the book. To use the plugins,
 you should run:
 
 ```bash
-$ cargo install --locked --path packages/mdbook-trpl-listing
-$ cargo install --locked --path packages/mdbook-trpl-note
+$ cargo install --locked --path packages/mdbook_trpl
 ```
 
 ## Building
