--- packages/mdbook-trpl-note/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "mdbook-trpl-note"
-version = "1.0.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4", features = ["derive"] }
-mdbook = { version = "0.4", default-features = false }     # only need the library
-pulldown-cmark = { version = "0.10", features = ["simd"] }
-pulldown-cmark-to-cmark = "13"
-serde_json = "1"
-
-[dev-dependencies]
-assert_cmd = "2"
-
-# This package is used as a path dependency in `rust-lang/rust`, not published
-# to crates.io, so it cannot be part of the `rust-lang/book` workspace, because
-# path dependencies do not get built as a crate within the hosting workspace.
-[workspace]
