+++ packages/mdbook-trpl/README.md
@@ -1,0 +1,13 @@
+# mdbook_trpl
+
+A shared package for [mdbook][mdbook] [preprocessors][pre] used in [_The Rust
+Programming Language_][trpl].
+
+Supplies the following preprocessor binaries:
+
+- [mdbook-trpl-note](./src/bin/note)
+- [mdbook-trpl-listing](./src/bin/listing)
+
+[mdbook]: https://crates.io/crates/mdbook
+[pre]: https://rust-lang.github.io/mdBook/format/configuration/preprocessors.html
+[trpl]: https://doc.rust-lang.org/book/
